fix: explicit cast in getRay and document Vup precondition

+++ src/Camera.h
@@ -19,6 +19,8 @@ public:
     Vetor U;   // unit_vector(cross(Vup,W)) — points right
     Vetor V;   // cross(W, U)               — corrected up (already unit)
 
+    // Precondition: Vup must not be parallel to (C - M).
+    // If Vup || W, cross(Vup, W) is the zero vector and U/V become NaN.
     Camera(Ponto C, Ponto M, Vetor Vup, double d, int hres, int vres)
         : C(C), M(M), Vup(Vup), d(d), hres(hres), vres(vres)
     {
@@ -30,8 +32,8 @@ public:
     ray getRay(int i, int j) const {
         Ponto screen_center = C + W * (-d);
         Ponto pixel_point = screen_center
-            + U * (-0.5 + (i + 0.5) / hres)
-            + V * ( 0.5 - (j + 0.5) / vres);
+            + U * (-0.5 + (i + 0.5) / static_cast<double>(hres))
+            + V * ( 0.5 - (j + 0.5) / static_cast<double>(vres));
         return ray(C, unit_vector(pixel_point - C));
     }
 };
